add request interceptop, change rest client to okhttp, fix entities and small refactor

+++ src/main/java/com/money_account_service/dtos/response/CreateAccountResponseDto.java
@@ -2,8 +2,6 @@ package com.money_account_service.dtos.response;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
-public record CreateAccountResponseDto(String accountNumber, BigDecimal balance, String currency) {
+public record CreateAccountResponseDto(String accountNumber, Long balance, String currency) {
 }
