+++ src/main/java/com/money_account_service/dtos/response/TopUpResponseDto.java
@@ -2,8 +2,6 @@ package com.money_account_service.dtos.response;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
-public record TopUpResponseDto(BigDecimal balance) {
+public record TopUpResponseDto(Long balance) {
 }
